API endpoints change, create api module
1. Make endpoints more restful
2. move endpoints and models into api module
TODO:
1. error handling in the room page
2. password hash function
3. CSS styling of the page
4. third-party login system
5. use .env for sensitive keys
6. deploy to the internet
