+++ frontend/src/services/UserServices.js
@@ -1,13 +1,8 @@
 import axios from "axios";
 
-const UserInRoom = () => {
-    const request = axios.post("/api/user-in-room");
-    return request.then(response => response.data);
-}
-
 const UserAuth = () => {
     const request = axios.get("/api/user/auth");
     return request.then(response => response.data);
 }
 
-export default {UserInRoom, UserAuth};
+export default {UserAuth};
